partner and admin apis are done

+++ app/Http/Controllers/SpecifyProductController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use App\Models\Store;
 use App\Models\FeatureStore;
+use App\Models\Feature;
 use App\Http\Requests\SpecifyProductRequest;
 use Illuminate\Support\Facades\DB;
 
@@ -19,10 +20,11 @@ class SpecifyProductController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
         $store = Auth::user()->store;
-        $features = $store->features;
+        $limit = $request->input('limit', 10);
+        $features = $store->features()->paginate($limit)->items();
         if (!$features){
             return response()->json([
             'message' => 'no features found for this store'
@@ -40,24 +42,26 @@ class SpecifyProductController extends Controller
      */
     public function show(FeatureStore $feature)
     {
-        // $store_id = Auth::user()->store->id;
-        // // $store = $user->store->id;
-
-        // $something = DB::table('feature_store')
-        // ->where('store_id', $store_id)
-        // ->where('feature_id', $feature->id)
-        // ->get();
-        // // ->load('designs');
-
-        // return $something;
+         $store = Auth::user()->store;
+        
         if (!$feature){
-            return response()->json([
-            'message' => 'feature not found'
-        ], 404);
-        } else
+            return response()->json([ 'message' => 'feature not found'], 404);
+        } 
+
+         if ($feature->store_id !== $store->id) {
+        return response()->json(['message' => 'This feature does not belong to your store.'], 403);
+         }
+
+       $featurestore = \App\Models\Feature::find($feature->feature_id);
+    $storeInfo = \App\Models\Store::find($feature->store_id);
+
+
         return response()->json([
             'message' => 'feature found',
-            'data' => $feature // Return the products in JSON format
+            'store feature' => $featurestore,
+            'feature' => $feature,
+            'store' => $storeInfo
+
         ], 200);
     }
     
@@ -70,22 +74,29 @@ class SpecifyProductController extends Controller
         $store = Auth::user()->store;
        
          $feature = $request->validated();
+         $featureId = $feature['id'];
 
-         //if($feature['id'] )
+    if ($store->features()->wherePivot('feature_id', $featureId)->exists()) {
 
-          if (!$store->features()->wherePivot('feature_id', $feature['id'])->exists()) {
-        $store->features()->attach($feature['id']);
+    $pivot = FeatureStore::withTrashed()
+        ->where('store_id', $store->id)
+        ->where('feature_id', $featureId)->first();
+
+        if ($pivot && $pivot->trashed()) {
+        $pivot->restore();
+        return response()->json(['message' => 'Feature restored successfully'], 200);
+    } else {
+        return response()->json(['message' => 'Feature is selected before'], 403);
     }
+} 
+        $store->features()->attach($featureId);
 
-   
+        $savedFeature = Feature::with('product_category')->find($featureId);
 
-    // $store->features()->attach(['feature_id'=>$feature['id']]); 
-    // $store->save(); 
-
-    return response()->json([
+        return response()->json([
         'message' => 'Feature selected successfully',
-        'data' => $store,
-    ],201);
+        'data' => $savedFeature,
+            ],201);
     }
 
 
@@ -96,10 +107,20 @@ class SpecifyProductController extends Controller
      */
     public function destroy(FeatureStore $feature)
     {
+         $store = Auth::user()->store;
+    
+        if (!$feature) {
+            return response()->json(['message' => 'Feature not found.'], 404);
+        }
+
+         if ($feature->store_id !== $store->id) {
+        return response()->json(['message' => 'This feature does not belong to your store.'], 403);
+         }
+
         $feature->delete();
 
         return response()->json([ // Return a JSON response indicating success
-            'message' => 'feature Deleted Successfully'
+            'message' => 'Feature Deleted Successfully'
         ], 200);
     }
 }
